style: rename course_review_model.py to review_model.py
